+++ ChangeLog.md
@@ -1,3 +1,9 @@
+# 1.6.7
+
+* flush stdout when using `echo`, not just `echo_n`
+* fix should be able to silence stderr when using `runHandle`
+* expose RunFailed
+
 # 1.6.6
 
 * add prependToPath function